\name{readme}
\alias{readme}

\title{readme}

\description{readme}

\usage{
readme.res <- readme(undergradlist = list(), trainingset = NULL, testset =
                   NULL, formula = NULL, features = NULL, n.subset =
                   NULL, prob.wt = NULL, boot.se = NULL, nboot = NULL, printit = NULL)
}

\arguments{
  \item{undergradlist}{undergrad...}
  \item{trainingset}{training set. ..}
  \item{testset}{test set ...}
  \item{formula}{formula ...}
  \item{features}{features ..}
  \item{n.subset}{n.subset ...}
  \item{prob.wt}{prob.wt ..}
  \item{boot.se}{boot.se ..}
  \item{nboot}{nboot ...}
  \item{printit}{printit ...}
}

\value{
  what it returns ?
}
